```
This message tells you that you are about to commit the changes to the `about.html` file. However, the `contact.html` file has been modified but not staged for commit. The `home.html` file has been deleted and is not staged for commit. The `stylingfile.css` file has been modified and is not staged for commit. The `.DS_Store` file is untracked and is not staged for commit. The `index.html` file is untracked and is not staged for commit.

After you have staged the changes you want to commit, you can save and close the editor. The changes will be committed to the repository.

yeah